tests: Re-enable snapshot
