+++ sdk/python/localcluster/snapshot.py
@@ -20,8 +20,8 @@ class Snapshot:
         self.parent_dir = parent_dir
         self.cluster = cluster
 
-    def create(self, anvil_file: Path):
-        logging.info("Taking snapshot")
+    def create(self):
+        logging.info(f"Taking snapshot (anvil port: {self.anvil_port}")
 
         # delete old snapshot
         shutil.rmtree(self.sdir, ignore_errors=True)
@@ -45,10 +45,8 @@ class Snapshot:
 
             db_target_dir.mkdir(parents=True, exist_ok=True)
 
-            # FIXME: This breaks the random base port approach since nodes will make announcements using different port.
-            # Skipping for now until a better approach is found.
-            # for file in EXPECTED_FILES_FOR_SNAPSHOT:
-            #   shutil.copy(source_dir.joinpath(file), db_target_dir)
+            for file in EXPECTED_FILES_FOR_SNAPSHOT:
+                shutil.copy(source_dir.joinpath(file), db_target_dir)
 
             shutil.copy(source_dir.joinpath("./hoprd.id"), target_dir)
             shutil.copy(source_dir.joinpath("./.env"), target_dir)
@@ -60,7 +58,7 @@ class Snapshot:
 
         for entry in self.parent_dir.glob("*"):
             if entry.is_dir():
-                if entry.name == "snapshot":
+                if entry.name == f"snapshot-{self.anvil_port}":
                     continue
                 shutil.rmtree(entry, ignore_errors=True)
             else:
@@ -78,11 +76,9 @@ class Snapshot:
             self.sdir.joinpath("default.cfg.yaml"),
         ]
 
-        # FIXME: This breaks the random base port approach since nodes will make announcements using different port.
-        # Skipping for now until a better approach is found.
-        # for i in range(self.cluster.size):
-        #     node_dir = self.sdir.joinpath(f"{NODE_NAME_PREFIX}_{i+1}")
-        #     expected_files.extend([node_dir.joinpath(file) for file in EXPECTED_FILES_FOR_SNAPSHOT])
+        for i in range(self.cluster.size):
+            node_dir = self.sdir.joinpath(f"{NODE_NAME_PREFIX}_{i+1}")
+            expected_files.extend([node_dir.joinpath(file) for file in EXPECTED_FILES_FOR_SNAPSHOT])
 
         for f in expected_files:
             if not f.exists():
@@ -93,4 +89,4 @@ class Snapshot:
 
     @property
     def sdir(self):
-        return self.parent_dir.joinpath("snapshot")
+        return self.parent_dir.joinpath(f"snapshot-{self.anvil_port}")
